rename base directory to compass
