--- src/vivarium_ciff_sam/__init__.py
@@ -1,5 +1,0 @@
-"""vivarium_ciff_sam
-
-Research repository for the vivarium_ciff_sam project.
-
-"""
