--- src/vivarium_ciff_sam/components/__init__.py
@@ -1,6 +1,0 @@
-from vivarium_ciff_sam.components.intervention import SQLNSIntervention, WastingTreatmentIntervention
-from vivarium_ciff_sam.components.observers import (CategoricalRiskObserver, DisabilityObserver, DiseaseObserver,
-                                                    MortalityObserver)
-from vivarium_ciff_sam.components.lbwsg import LowBirthWeight, ShortGestation
-from vivarium_ciff_sam.components.treatment import SQLNSTreatment
-from vivarium_ciff_sam.components.wasting import ChildWasting
