+++ src/vivarium_compass_sam/__about__.py
@@ -3,8 +3,8 @@ __all__ = [
     "__email__", "__license__", "__copyright__",
 ]
 
-__title__ = "vivarium_ciff_sam"
-__summary__ = "Research repository for the vivarium_ciff_sam project."
+__title__ = "vivarium_compass_sam"
+__summary__ = "Research repository for the vivarium_compass_sam project."
 __uri__ = "https://github.com/ihmeuw/vivarium_ciff_sam"
 
 __version__ = "1.0.0"
