+++ src/vivarium_compass_sam/__init__.py
@@ -1,0 +1,5 @@
+"""vivarium_compass_sam
+
+Research repository for the vivarium_compass_sam project.
+
+"""
