+++ src/vivarium_compass_sam/tools/make_results.py
@@ -3,7 +3,7 @@ import shutil
 
 from loguru import logger
 
-from vivarium_ciff_sam.results_processing import process_results
+from vivarium_compass_sam.results_processing import process_results
 
 
 def build_results(output_file: str, single_run: bool):
